LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-prepended-null.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Connection should fail immediately, rather than succeeding or staying in limbo until timeout, if a null byte is received before the handshake.');

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

var timedOut = false;
var connected = false;
var origin;

function endTest() {
    shouldBeFalse('timedOut');
    shouldBeFalse('connected');
    shouldBeUndefined('origin');
    clearTimeout(timeoutID);
    finishJSTest();
}

var url = 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-prepended-null';
var ws = new WebSocket(url);

ws.onopen = function()
{
    debug('Connected');
    connected = true;
}

ws.onmessage = function(messageEvent)
{
    origin = messageEvent.data;
    debug('origin = ' + origin);
    ws.close();
}

ws.onclose = function()
{
    endTest();
}

function timeoutCallback()
{
    debug('Timed out (state = ' + ws.readyState + ')');
    timedOut = true;
    endTest();
}

var timeoutID = setTimeout(timeoutCallback, 3000);

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
