LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-prepended-null-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-prepended-null' failed: Status line contains embedded null
Connection should fail immediately, rather than succeeding or staying in limbo until timeout, if a null byte is received before the handshake.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS timedOut is false
PASS connected is false
PASS origin is undefined.
PASS successfullyParsed is true

TEST COMPLETE
